agentcore/runtime/pyproject.toml
[project]
name = "ocr-agent-runtime"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "boto3>=1.35.0",
    "strands-agents>=1.0",
    "aws-opentelemetry-distro>=0.10.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
